Modify header
