+++ generic_type_support/src/examples/check.cpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "generic_type_support/generic_type_support.hpp"
-
+#include <generic_type_support/generic_type_support.hpp>
 #include <iostream>
 
 int main()
