+++ generic_type_support/src/examples/print.cpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "generic_type_support/generic_type_support.hpp"
-
+#include <generic_type_support/generic_type_support.hpp>
 #include <iostream>
 #include <rclcpp/rclcpp.hpp>
