+++ generic_type_support/src/impl/message.cpp
@@ -14,7 +14,7 @@
 
 #include "message.hpp"
 #include "field.hpp"
-#include "generic_type_support/errors.hpp"
+#include <generic_type_support/errors.hpp>
 #include <iostream>
 
 namespace generic_type_support
